refactor: separate module for SPARQL update
